projects/api/pyproject.toml
[tool.poetry]
name = "api-server"
version = "0.1.0"
description = "FastAPI server for Algorand hackathon"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.0.0"
py-algorand-sdk = "^2.6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"